requirements.txt
--index-url https://mirrors.aliyun.com/pypi/simple/


langchain
langchain-core
langchain-openai
langgraph
pandas
matplotlib
numpy
python-dotenv
# LLM Observability
langfuse
# Additional utilities for observability
pydantic>=2.0.0
pyyaml>=6.0
httpx
fastapi
uvicorn 